Fix MIME type registration
Fixes #9.

It seems like the TryExec key was what prevented the MIME type
association from working as intended. By removing it, the
"default association" works fine.

The xdg-mime hacks never worked, and can be removed, as the real work
is performed by the dpkg hooks of desktop-file-utils, shared-mime-info
etc. Therefore, they are removed.