Switch to ~/Applications
For existing users, the transition will be performed gracefully by
notifying them about the change and asking them to do it before
moving any files.

The code is prepared for future changes, which hopefully won't be
required (unless ~/Applications becomes a standard, and on non-English
desktops the files need to be moved into some other directory).