Add fallback solution for older systems
glibc prior to 2.27 doesn't provide memfd_create. Ubuntu xenial for instance ships with a libc6 package containing an older version of glibc. Therefore, we need a fallback.

This commit introduces a similar approach, based on the inferior shm_open functionality. It works fine and is reasonably safe, but not as rock solid as memfd_create (much due to the fact that we cannot use O_EXCL, as we cannot use a writable file descriptor with fexecve(...)).