Create unprivileged user in Docker images again
The docker build script enforces the use of the caller's uid inside the container (works well, even though there's no entry in the image's /etc/passwd). However, when using the container without the script (e.g., for debugging), root is used by default now. This is not a good idea. Therefore, we just create *some* user and use it by default. This doesn't add a dependency on the caller's UID, unlike the previous version.